+++ client/changelog.md
@@ -1,2 +1,3 @@
 ### Client
 - Fixed crackling/noise in microphone capture on Minecraft 26.1.x.
+- Fixed voice distance sphere rendering on join when it shouldn't.